requirements.txt
selenium
telethon
python-dotenv
requests
beautifulsoup4
thefuzz
python-Levenshtein
tqdm
Flask